+++ application/models/FleetInfo.php
@@ -29,7 +29,16 @@ class FleetInfo extends CSV_Model
         return $tmp;
 
     }
+    
+    // provide form validation rules
+    public function rules() {
+        $config = array(
+            ['field' => 'airid', 'label' => 'Plane ID', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+        );
+        return $config;
+    }
 
+    
 
     // retrieve a single plane, null if not found