+++ application/views/fleetedit.php
@@ -1,0 +1,8 @@
+<h1>{modifyTitle}</h1>
+<form role="form" action="/fleet/submit" method="post">
+    {fairid}<br>
+    {zsubmit}<br>
+</form>
+{error}
+<a href="/fleet/cancel"><input type="button" value="Cancel"/></a>
+<a href="/fleet/delete"><input type="button" value="Delete"/></a>